AI/requirements.txt
fastapi
uvicorn
spacy
langchain_google_genai
langchain
python-dotenv